Long userId

+++ src/main/java/se/anyro/tgbotapi/types/Contact.java
@@ -7,6 +7,6 @@ public class Contact {
     public String phone_number;
     public String first_name;
     public String last_name;
-    public int user_id;
+    public long user_id;
     public String vcard;
 }
